requirements.txt
aiogram==2.14.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
crispy-bootstrap5==2025.6
Django==4.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environs==14.2.0
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
marshmallow==4.0.0
multidict==6.4.4
packaging==25.0
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
sqlparse==0.5.3
yarl==1.20.0
